backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
langchain-openai
sse-starlette
pydantic
langchain-community